CASTS and etc. to try to get this code to compile.  It still won't, but this is closer.

+++ zkdecrypto/headers/command.h
@@ -380,7 +380,7 @@ inline int CommandSolve(int cmd_id)
 			if(DialogBox(hInst,MAKEINTRESOURCE(IDD_NUMBER),hMainWnd,(DLGPROC)NumberProc))
 			{
 				if(iNumber<2) iNumber=2;
-				iSqrt=sqrt(iNumber);
+				iSqrt=sqrt((double)iNumber);
 				
 				for(cur_fact=iNumber; cur_fact>iSqrt; cur_fact--)
 					if(!(iNumber%cur_fact))
